added comments
normal operation and request operation works through one cycle. Then when you try to read a file again through client, errSim will receive the packet, send it, and hang. Possible reason being the errSim saves the port number of the current connection, and when the server finishes sending the file the connection is closed and that port number is no longer used. Need a way for the errorSimulator to find out when the file is completely sent so it knows it needs to update serverPort1 variable.